fixed function and swiper
