+++ src/partials/reviews.html
@@ -1,7 +1,7 @@
 <section class="reviews">
     <div class="container">
         <h2 class="reviews__heading">REVIEWS</h2>
-        <div class="swiper">
+        <div class="swiper reviews-swiper">
             <ul class="reviews__list swiper-wrapper js-reviews"></ul>
             <div class="reviews__buttons js-reviews-btn">
                 <button class="reviews__btn reviews__btn--left reviews__btn--prev js-swiper-prev" type="button">
